Created McAffee-Async.py
In this version of the script, we have converted the export_mcafee_info function to an asynchronous function using the async keyword. We have also defined a list of McAfee Mail Gateway devices, each with a hostname, username, and password. The main function uses asyncio.gather to run the export_mcafee_info function for each device asynchronously and wait for all tasks to complete. Finally, the asyncio.run function is called to start the event loop and run the main function.

This script allows you to export information from multiple McAfee Mail Gateway devices at the same time, making the process faster and more efficient.